feat(toolchain-updates): concurrently execute each toolchain's update
This change implied the introduction of the "futures_util" crate.
Feedback is greately appreciated on the above.

To ensure that there are no hangs, a special "if" was needed to
guarantee that we never try running `.buffered()` with 0 (zero).

PS: The test `suite::cli_rustup::rustup_no_channels` shows the absence of
the hang as it tries to run `.buffered()` with 0.

Kudos to @rami3l for helping me figure out why this test was hanging.