feat(check-updates): concurrently check channels for updates (WIP)
This is still a work in progress.

Currently, to avoid race conditions of futures writing to the terminal,
the output is only being displayed after every channel is checked.

As this goes against the current behavior of the command, an alternative
with `indicatif` is being investigated to ensure that we can print
channels (in order) as they are being checked for updates.

Benchmarking (100 runs) showed that this concurrent approach is about
2.2 times faster than the sequential approach.

NB: This commit was done to solely show the progress of the
implementation, it is not *production-ready*.
NB2: The results of the benchmark may be impacted by the introduction of
`indicatif`.