feat(updates): check for updates concurrently
This change implied the introduction of the "futures_util" crate.
Feedback is greately appreciated on the above.

To ensure that there are no hands, a special "if" was needed to
guarantee that we never try running `.buffered()` with 0 (zero).
Otherwise, it would hang (kudos to rami3l for helping me out with
this).

To report progress in a concurrent way whilst maintaining order, the
`indicatif` crate was used.

This commit will be followed by a patch that aims to put the progress
bars correctly writing to stdout (with attention to ANSI escape codes).