feat(updates): check for updates concurrently
This change implied the introduction of the "futures_util" crate.
Feedback is greately appreciated on the above.

To ensure that there are no hangs, a special "if" was needed to
guarantee that we never try running `.buffered()` with 0 (zero).
Otherwise, it would hang (kudos to rami3l for helping me out with
this).
https://github.com/rust-lang/futures-rs/pull/1194#discussion_r209501774

To report progress in a concurrent way whilst maintaining order, the
`indicatif` crate was used.