Implement the `TermLike` trait for `ColorableTerminal`
If we are not in a *tty* the progress will be hidden and will instead be
printed after every channel is checked for updates.

NB: since we are using `indicatif` to display progress, when running in
a *tty*, we can use `buffer_unordered` instead of `buffered`, as
`indicatif` already handles the output in the correct order.
A quick benchmark (100 runs + 10 warmpu runs in a 50Mbps connection)
showed that this small change yields a **1.2x** speedup.