Implement the `TraitLike` trait for `ColorableTerminal`
The `ColorableTerminal` now has the `TermLike` trait.
If we are not in a *tty* the progress will be hidden and will instead be
printed after every channel is checked for updates.

NB: since we are using `indicatif` to display progress, when running in
a *tty*, we can use `buffer_unordered` instead of `buffered`, as
`indicatif` already handles the output in the correct order.